Orbital Improvements and Return I've added parameters to each script to make them more versatile.
I've renamed launchComSat to just launchSat to make it a less 
specialized script. This launches a craft and before circularizing it 
stages away any excess stages.

returnOrbit now relies on staging to deploy a parachute instead of 
chutessafe. The option chutessafe wasn't working for me so instead I 
stage the parachute, which if set properly cannot deploy when in unsafe 
conditions.

Todo: Compensate for apoapsis changes during circularization burns.